% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{AddCumulativeRow}
\alias{AddCumulativeRow}
\title{AddCumulativeRow}
\usage{
AddCumulativeRow(DataFrame, StringInNonNumericColumns = "All")
}
\description{
Adds a cumulative row to a tibble/data.frame
}
\keyword{internal}
