% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCPhosAlphaPalettes}
\alias{CCPhosAlphaPalettes}
\title{Alpha palettes used in CCPhos packages}
\format{
\subsection{\code{CCPhosAlphaPalettes}}{

A list of named vectors
\describe{
\item{name}{Palette name}
\item{values}{Vector of alpha values}
}
}
}
\source{
Own preferences
}
\usage{
CCPhosAlphaPalettes
}
\description{
A list of named vectors containing alpha values
}
\author{
Bastian Reiter
}
\keyword{datasets}
