% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{CheckDSConnections}
\alias{CheckDSConnections}
\title{CheckDSConnections}
\usage{
CheckDSConnections(DSConnections)
}
\arguments{
\item{DSConnections}{Usually a \code{list} of \code{DSConnection} class objects. If this is set to \code{NULL} this function will try to find an existing \code{list} programmatically and return it.}
}
\description{
This function checks if 'DSConnections' that are passed to a DataSHIELD client function are valid. If this object is not passed, the function will try to find an existing \code{list} programmatically (using \code{DSI} functionality) and return it.
}
\keyword{internal}
