% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{ComputeSurvCI}
\alias{ComputeSurvCI}
\title{ComputeSurvCI}
\usage{
ComputeSurvCI(SurvEstimate, SurvStdErr, Method = "Cloglog", ConfLevel = 0.95)
}
\arguments{
\item{SurvEstimate}{\code{numeric vector} - The timepoint-specific estimate for cumulative survival (Kaplan-Meier)}

\item{Method}{\code{string} - The method used for computing confidence interval limits. One of 'NormalApproximation' or 'Cloglog' (default).}

\item{ConfLevel}{\code{numeric scalar} - Chosen confidence level}

\item{SurvStdError}{\code{integer vector} - The timepoint-specific number of events in a life table}
}
\description{
Compute confidence interval limits for cumulative survival estimator. Can also be used for CI of cumulative hazard.
}
