% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{ComputeSurvHaz}
\alias{ComputeSurvHaz}
\title{ComputeSurvHaz}
\usage{
ComputeSurvHaz(n.risk, n.event, CumHazMethod = "NelsonAalen")
}
\arguments{
\item{n.risk}{\code{integer vector} - The timepoint-specific number at risk in a life table}

\item{n.event}{\code{integer vector} - The timepoint-specific number of events in a life table}

\item{CumHazMethod}{\code{string} - The method used for cumulative hazard estimation. One of 'KaplanMeier' or 'NelsonAalen' (default).}
}
\description{
Compute estimates (with standard errors) for cumulative survival and cumulative hazard based on timepoint-specific number at risk at number of events from a life table.
}
