% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplayTimeSeries.R
\name{DisplayTimeSeries}
\alias{DisplayTimeSeries}
\title{DisplayTimeSeries}
\usage{
DisplayTimeSeries(
  TimeSeriesData,
  TimePointFeature,
  ValueFeature,
  GroupingFeature = NULL,
  IncludeMissingTimePoints = FALSE
)
}
\arguments{
\item{TimeSeriesData}{\code{data.frame} - Containing data associated with points in time}

\item{TimePointFeature}{\code{symbol} or \code{string} - Feature in 'TimeSeriesData' that contains time points (e.g. years)}

\item{ValueFeature}{\code{symbol} or \code{string} - Feature in 'TimeSeriesData' that contains values}

\item{GroupingFeature}{\code{symbol} or \code{string} - Optional feature in 'TimeSeriesData' that contains grouping variable}

\item{IncludeMissingTimePoints}{\code{logical} - Indicates whether discrete time points (e.g. years) with no data should be displayed in the time series - Default: \code{FALSE}}
}
\value{
A \code{data.frame} displaying time series data in horizontal form
}
\description{
Transform time series data to display-friendly format with time points as column names.
}
\author{
Bastian Reiter
}
