% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CreateSurvObject.R
\name{ds.CreateSurvObject}
\alias{ds.CreateSurvObject}
\title{ds.CreateSurvObject}
\usage{
ds.CreateSurvObject(
  TableName,
  TimeFeature,
  EventFeature,
  MinFollowUpTime = 1,
  OutputName = "SurvObject",
  DSConnections = NULL
)
}
\arguments{
\item{TableName}{\code{string} - Name of the table containing the features of concern}

\item{TimeFeature}{\code{string} - Name of time feature}

\item{EventFeature}{\code{string} - Name of event feature}

\item{MinFollowUpTime}{\code{integer} - Optional minimum of observed follow up time}

\item{OutputName}{\code{string} - Assigned symbol name on servers}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list}
}
\description{
Create a \code{survival::Surv()} object on servers
}
\details{
Linked to server-side \code{ASSIGN} function \code{CreateSurvObjectDS()}.
}
\author{
Bastian Reiter
}
