% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.ExtractFromList.R
\name{ds.ExtractFromList}
\alias{ds.ExtractFromList}
\title{ds.ExtractFromList}
\usage{
ds.ExtractFromList(
  ListName,
  ObjectName,
  NewObjectName = NULL,
  DSConnections = NULL
)
}
\arguments{
\item{ListName}{\code{string} - Name of a \code{list} object on server}

\item{ObjectName}{\code{string} - Name of object inside \code{list}}

\item{NewObjectName}{\code{string} - Optionally assigned name of object after extraction to server session - Default: \code{NULL}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Auxiliary function to trigger assignment of object extracted from a list on server
}
\details{
Linked to server-side ASSIGN method ExtractFromListDS()
}
\author{
Bastian Reiter
}
