% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetCrossTab.R
\name{ds.GetCrossTab}
\alias{ds.GetCrossTab}
\title{ds.GetCrossTab}
\usage{
ds.GetCrossTab(TableName, FeatureNames, RemoveNA = FALSE, DSConnections = NULL)
}
\arguments{
\item{TableName}{\code{string} - Name of \code{data.frame} on server}

\item{FeatureNames}{\code{character} - Vector of feature names}

\item{RemoveNA}{\code{logical} - Indicating whether missing values should be removed prior to cross tabulation - Default: \code{FALSE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list}
\itemize{ \item CrossTab (\code{list}),
\item ChiSquaredTest (\code{list}) }
}
\description{
Perform cross tabulation for an arbitrary number of features in a table on the servers.
}
\details{
Linked to server-side \code{AGGREGATE} function \code{GetCrossTabDS()}
}
\author{
Bastian Reiter
}
