% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetFrequencyTable.R
\name{ds.GetFrequencyTable}
\alias{ds.GetFrequencyTable}
\title{ds.GetFrequencyTable}
\usage{
ds.GetFrequencyTable(
  TableName,
  FeatureName,
  GroupingFeatureName = NULL,
  MaxNumberCategories = 10,
  RemoveNA = FALSE,
  DSConnections = NULL
)
}
\arguments{
\item{TableName}{\code{string} - Name of the table containing the feature of concern}

\item{FeatureName}{\code{string} - Name of feature}

\item{GroupingFeatureName}{\code{string} - Name of optional grouping feature from the same table}

\item{MaxNumberCategories}{\code{integer} - Maximum number of categories analyzed individually before frequencies are cumulated in 'Other' category. - Default: 10}

\item{RemoveNA}{\code{logical} - Indicating whether missing values should be removed prior to frequency calculation - Default: \code{FALSE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} containing:
\itemize{\item AbsoluteFrequencies (\code{tibble}: Absolute value frequencies)
\item RelativeFrequencies (\code{tibble}: Relative value frequencies)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Get table of absolute and relative value frequencies for a nominal / ordinal feature.
}
\details{
Linked to server-side \code{AGGREGATE} function \code{GetFrequencyTableDS()}.
}
\author{
Bastian Reiter
}
