% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetTableCheck.R
\name{ds.GetTableCheck}
\alias{ds.GetTableCheck}
\title{ds.GetTableCheck}
\usage{
ds.GetTableCheck(
  TableName,
  RequiredFeatureNames = NULL,
  GetTemplate = FALSE,
  DSConnections = NULL
)
}
\arguments{
\item{TableName}{\code{string} - Name of \code{data.frame} or \code{tibble}}

\item{RequiredFeatureNames}{\code{character vector} - Optional names of required features - Default: \code{names()} applied to Table evaluated from \code{TableName}}

\item{GetTemplate}{\code{logical} - If set to \code{TRUE}, the function returns a template incorporating required feature names without actually checking an existing table}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment. Default: \code{FALSE}}
}
\value{
A \code{list} containing compiled meta data about table:
\itemize{\item TableCheckOverview
\item FeatureExistence
\item FeatureTypes
\item NonMissingValueCounts
\item NonMissingValueRates }
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Check out a table (\code{data.frame} or \code{tibble}) on servers and return informative meta data.
}
\details{
Linked to server-side AGGREGATE method \code{GetTableCheckDS()}
}
\author{
Bastian Reiter
}
