% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetTestData.R
\name{ds.GetTestData}
\alias{ds.GetTestData}
\title{ds.GetTestData}
\usage{
ds.GetTestData(
  DataSetName = NA_character_,
  TableName = NA_character_,
  SampleSize,
  Shuffle = TRUE,
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{DataSetName}{\code{string} - Name of Data Set object (list) on server}

\item{TableName}{\code{string} - Name of specific table object on server}

\item{SampleSize}{\code{count} - Sample size of test data}

\item{Shuffle}{\code{flag} - Whether to shuffle test data}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{flag} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
A \code{list}
}
\description{
Trigger basic preparatory transformations on RawDataSet prior to Curation:
\itemize{  \item Feature name harmonization
\item Add ID feature (running number) for tables without primary key feature
\item Type- or format-conversion of features }
}
\details{
Linked to server-side ASSIGN method \code{PrepareRawDataDS()}
}
\author{
Bastian Reiter
}
