% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.JoinTables.R
\name{ds.JoinTables}
\alias{ds.JoinTables}
\title{ds.JoinTables}
\usage{
ds.JoinTables(
  TableNameA,
  TableNameB,
  ByStatement,
  JoinType = "left_join",
  OutputName,
  DSConnections = NULL
)
}
\arguments{
\item{TableNameA}{\code{string} - Name of Table A on server}

\item{TableNameB}{\code{string} - Name of Table B on server}

\item{ByStatement}{\code{string} - The insides of a \code{dplyr::join_by()}-Statement defining how to join tables}

\item{JoinType}{\code{string} - Name of \code{dplyr::join}-function used, one of:
\itemize{\item 'left_join' (Default)
\item 'right_join'
\item 'full_join'
\item 'inner_join'}}

\item{OutputName}{\code{string} - Name of resulting table on server}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Join tables on server, making use of \code{dplyr} mutating join operations.
}
\details{
Linked to server-side \code{ASSIGN} function \code{JoinTablesDS()}
}
\author{
Bastian Reiter
}
