% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.MakeList.R
\name{ds.MakeList}
\alias{ds.MakeList}
\title{ds.MakeList}
\usage{
ds.MakeList(ObjectNames, OutputName, DSConnections = NULL)
}
\arguments{
\item{ObjectNames}{\code{character} - Names of objects on server to be bundled in a list. Can be a \code{named vector} if names in list should differ from symbol names (with names in named vector being the names in list).}

\item{OutputName}{\code{string} - Name of resulting \code{list} on server}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Bundle objects in a list on servers.
}
\details{
Linked to server-side \code{ASSIGN} function \code{MakeListDS()}
}
\author{
Bastian Reiter
}
