% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.MutateTable.R
\name{ds.MutateTable}
\alias{ds.MutateTable}
\title{ds.MutateTable}
\usage{
ds.MutateTable(
  TableName,
  MutateExpression,
  GroupBy = NULL,
  OutputName,
  DSConnections = NULL
)
}
\arguments{
\item{TableName}{\code{string} - Name of \code{data.frame} on server}

\item{MutateExpression}{\code{string} - \code{dplyr::mutate} expression as string}

\item{GroupBy}{\code{string} - Optional \code{dplyr::group_by} expression as string}

\item{OutputName}{\code{string} - Name of resulting \code{data.frame} on server}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Make use of \code{dplyr::mutate()} to create new features in a given table (\code{data.frame}) on the servers.
}
\details{
Linked to server-side \code{ASSIGN} function \code{MutateTableDS()}
}
\author{
Bastian Reiter
}
