% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsFredaClient-package.R
\docType{package}
\name{dsFredaClient-package}
\alias{dsFredaClient}
\alias{dsFredaClient-package}
\title{dsFredaClient: Federated Refinement, Exploration and Data Augmentation}
\description{
Client-side DataSHIELD R package as part of a Framework for Federated Data Processing
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ccphos.netlify.app/}
  \item \url{https://github.com/BastianReiter/dsFredaClient}
}

}
\author{
\strong{Maintainer}: Bastian Reiter \email{b.reiter@med.uni-frankfurt.de} (\href{https://orcid.org/0000-0002-1088-0502}{ORCID})

}
\keyword{internal}
