% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreFeature.R
\name{ExploreFeature}
\alias{ExploreFeature}
\title{ExploreFeature}
\usage{
ExploreFeature(
  TableName,
  FeatureName,
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async,
  ...
)
}
\arguments{
\item{TableName}{\code{string} - Name of the table containing the feature of concern}

\item{FeatureName}{\code{string} - Name of feature}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
\code{list}
\itemize{ \item FeatureInfo
\item Statistics }
}
\description{
Get characterizing statistics about a feature of arbitrary data type.
}
\author{
Bastian Reiter
}
