% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportPlot.R
\name{ExportPlot}
\alias{ExportPlot}
\title{ExportPlot}
\usage{
ExportPlot(
  Plot,
  Directory,
  FileName = "default",
  FileFormat = "svg",
  Width,
  Height,
  Unit = "cm",
  DPI = "print",
  LegendPosition = NULL,
  ShowAxisTitle_x = TRUE,
  ShowAxisTitle_y = TRUE,
  ShowAxisLabels_x = TRUE,
  ShowAxisLabels_y = TRUE
)
}
\arguments{
\item{Plot}{A ggplot2 object}

\item{Directory}{Output directory}

\item{FileName}{Name of the created file (optional) -- Default: Take name from ggplot2 object}

\item{FileFormat}{File format (optional) -- Default: SVG -- Options: See ggplot2::ggsave() documentation}

\item{Width}{Plot Width in units passed by "Unit"}

\item{Height}{Plot Height in units passed by "Unit"}

\item{Unit}{Unit for size measures (optional) -- Default: cm}

\item{DPI}{Resolution in dot per inches (optional) -- Default: "print"}

\item{LegendPosition}{Legend position (optional) -- Default: NULL / no legend}

\item{ShowAxisTitle_x}{Whether to print X axis title}

\item{ShowAxisTitle_y}{Whether to print Y axis title}

\item{ShowAxisLabels_x}{Whether to print X axis labels}

\item{ShowAxisLabels_y}{Whether to print X axis labels}
}
\value{
No return; The function creates a plot file
}
\description{
Helper function to save \code{ggplot2} objects into graphic files, based on \code{ggplot2::ggsave()}
}
\author{
Bastian Reiter
}
