% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FredaAlphaPalettes}
\alias{FredaAlphaPalettes}
\title{Alpha palettes used in Freda packages}
\format{
\subsection{\code{FredaAlphaPalettes}}{

A list of named vectors
\describe{
\item{name}{Palette name}
\item{values}{Vector of alpha values}
}
}
}
\source{
Own preferences
}
\usage{
FredaAlphaPalettes
}
\description{
A list of named vectors containing alpha values
}
\author{
Bastian Reiter
}
\keyword{datasets}
