% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCollectiveExploration.R
\name{GetCollectiveExploration}
\alias{GetCollectiveExploration}
\title{GetCollectiveExploration}
\usage{
GetCollectiveExploration(
  OrderList = NULL,
  InputWorkspaceInfo = NULL,
  DSConnections = NULL
)
}
\arguments{
\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} containing overview and details of server-side workspace objects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
}
\author{
Bastian Reiter
}
