% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEligibleValues.R
\name{GetEligibleValues}
\alias{GetEligibleValues}
\title{GetEligibleValues}
\usage{
GetEligibleValues(TableName, FeatureName, Stage = "Curated")
}
\arguments{
\item{TableName}{\code{string} - Name of the table containing the feature of concern}

\item{FeatureName}{\code{string} - Name of feature}

\item{Stage}{\code{string} - Either 'Curated' (Default) or 'Raw'. Both stages might have different sets of eligible values.}
}
\value{
\code{vector} of eligible values
}
\description{
Get a set of eligible values for a feature, if defined in meta data.
}
\author{
Bastian Reiter
}
