% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetServerOpalDBInfo.R
\name{GetServerOpalDBInfo}
\alias{GetServerOpalDBInfo}
\title{GetServerOpalDBInfo}
\usage{
GetServerOpalDBInfo(
  ServerSpecifications = NULL,
  OpalTableNames.Required = NULL,
  OpalTableNames.Dictionary = NULL,
  DSConnections = NULL
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} - Same data frame used for login. Used here only for akquisition of server-specific project names (in case they are differing). - Default: NULL for virtual project}

\item{OpalTableNames.Required}{\code{character} - The table names expected/required in server Opal data base}

\item{OpalTableNames.Dictionary}{Optional \code{list} of named \code{character vectors} - To enable server-specific mapping of deviating to required Opal data base table names. Names of list elements must match server names. For rules that should be applied on all servers, choose form \code{list(All = c('LookupName' = 'RequiredName'))}.}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list}:
\itemize{ \item OpalTables.Available (\code{tibble})
\item OpalTables.Required (\code{tibble})
\item Summary (\code{tibble}) }
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Check if tables are available in server Opal data bases.
}
\author{
Bastian Reiter
}
