% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetServerOpalInfo.R
\name{GetServerOpalInfo}
\alias{GetServerOpalInfo}
\title{GetServerOpalInfo}
\usage{
GetServerOpalInfo(
  ServerSpecifications = NULL,
  RequiredTableNames,
  DSConnections = NULL
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} - Same data frame used for login. Used here only for akquisition of server-specific project names (in case they are differing). - Default: NULL for virtual project}

\item{RequiredTableNames}{\code{character vector} - The table names expected/required in server Opal data base}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{tibble}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Check if tables are available in server Opal data bases.
}
\author{
Bastian Reiter
}
