% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetServerWorkspaceInfo.R
\name{GetServerWorkspaceInfo}
\alias{GetServerWorkspaceInfo}
\title{GetServerWorkspaceInfo}
\usage{
GetServerWorkspaceInfo(
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: Taken from \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
A \code{list} containing overview and details of server-side workspace objects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Check which objects live in server-side R sessions and collect meta data about them.
}
\author{
Bastian Reiter
}
