% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTheme.R
\name{ggTheme}
\alias{ggTheme}
\title{ggTheme}
\usage{
ggTheme(
  ...,
  BaseSize = 11,
  LegendPosition = "right",
  SizeFactorPlotTitle = 1.67,
  SizeFactorPlotSubtitle = 1.33,
  SizeFactorPlotCaption = 1,
  SizeFactorAxisLabels = 1.33,
  SizeFactorTickLabels_x = 1,
  SizeFactorTickLabels_y = 1,
  SizeFactorLegendLabels = 1,
  SizeFactorFacetLabels = 1.33
)
}
\arguments{
\item{BaseSize}{Font size that relatively determines all other sizes}

\item{LegendPosition}{Legend position ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{SizeFactorPlotTitle}{Determines font size of plot title}

\item{SizeFactorPlotSubtitle}{Determines font size of plot subtitle}

\item{SizeFactorPlotCaption}{Determines font size of plot caption}

\item{SizeFactorAxisLabels}{Determines font size of main axis labels}

\item{SizeFactorTickLabels_x}{Determines font size of x axis tick labels}

\item{SizeFactorTickLabels_y}{Determines font size of y axis tick labels}

\item{SizeFactorLegendLabels}{Determines font size of legend labels}

\item{SizeFactorFacetLabels}{Determines font size of facet labels}
}
\description{
Custom ggplot2 theme, based on ggplot2::theme_classic()
}
\author{
Bastian Reiter
}
