% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtTheme.R
\name{gtTheme}
\alias{gtTheme}
\title{gtTheme}
\usage{
gtTheme(
  gtObject,
  Fontname = "sans",
  ShowNAs = FALSE,
  TableWidth = NULL,
  TableAlign = "center",
  BaseFontSize = "80\%",
  ...
)
}
\arguments{
\item{gtObject}{A gt table object}

\item{Fontname}{The name of the font to be used - Default: "sans" / If package \code{sysfonts} is installed, the Google font 'Karla' is chosen}

\item{ShowNAs}{logical; whether NA values should be printed}

\item{TableWidth}{\code{numeric}}

\item{TableAlign}{\code{string} - One of "center" / "left" / "right"}

\item{...}{\code{list} of arguments}
}
\value{
A (modified) gt table object
}
\description{
Custom gt theme
}
\details{
Can be applied to gt table objects from gt package
}
\author{
Bastian Reiter
}
