% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtTheme_CCP.R
\name{gtTheme_CCP}
\alias{gtTheme_CCP}
\title{gtTheme_CCP}
\usage{
gtTheme_CCP(
  gtObject,
  ShowNAs = FALSE,
  TableWidth = NULL,
  TableAlign = "center",
  BaseFontSize = "80\%",
  ...
)
}
\arguments{
\item{gtObject}{A gt table object}

\item{ShowNAs}{logical; whether NA values should be printed}

\item{...}{}
}
\value{
A (modified) gt table object
}
\description{
Custom gt theme
}
\details{
Can be applied to gt table objects from gt package
}
\author{
Bastian Reiter
}
