% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCurationReport.R
\name{MakeCurationReport}
\alias{MakeCurationReport}
\title{MakeCurationReport}
\usage{
MakeCurationReport(
  CurationReportData,
  PathToReportTemplate = "./Development/Reporting/CurationReport.qmd"
)
}
\arguments{
\item{CurationReportData}{List | Curation output object supplied by \code{\link{GetCurationReport}}}

\item{PathToReportTemplate}{String | Path to a qmd-file serving as template for a Quarto document to be rendered}
}
\description{
Create html output based on curation report objects supplied by \code{\link{GetCurationReport}}. Requires Quarto.
}
\author{
Bastian Reiter
}
