% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{MakeFunctionMessage}
\alias{MakeFunctionMessage}
\title{MakeFunctionMessage}
\usage{
MakeFunctionMessage(
  Text,
  IsClassInfo = FALSE,
  IsClassSuccess = FALSE,
  IsClassWarning = FALSE,
  IsClassFailure = FALSE
)
}
\arguments{
\item{Text}{\code{character}}

\item{IsClassInfo}{\code{logical}}

\item{IsClassSuccess}{\code{logical}}

\item{IsClassWarning}{\code{logical}}

\item{IsClassFailure}{\code{logical}}
}
\description{
Turn function message into named vector to enable classification of feedback
}
\keyword{internal}
