% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_ServerRequirements}
\alias{Meta_ServerRequirements}
\title{For evaluation of CCP server requirements}
\format{
An object of class \code{list} of length 2.
}
\usage{
Meta_ServerRequirements
}
\description{
List containing defined server requirements concerning package and function availability
}
\author{
Bastian Reiter
}
\keyword{datasets}
