% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyDSConnections.R
\name{ModifyConnections}
\alias{ModifyConnections}
\title{ModifyDSConnections}
\usage{
ModifyConnections(ServersToBeRemoved = NULL, DSConnections = NULL)
}
\arguments{
\item{ServersToBeRemoved}{\code{character vector} of server names to be removed from \code{DSConnections}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A (modified) \code{list} of \code{DSConnection} objects
}
\description{
Modifies given \code{list} of \code{DSConnection} objects created by DataSHIELD functionality.
}
\author{
Bastian Reiter
}
