% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{PrintMessages}
\alias{PrintMessages}
\title{PrintMessages}
\usage{
PrintMessages(Messages)
}
\arguments{
\item{Messages}{\code{list} List of of named vectors}
}
\description{
Take list of messages and print them with \code{PrintSoloMessage()}.
}
