% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{PrintSoloMessage}
\alias{PrintSoloMessage}
\title{PrintSoloMessage}
\usage{
PrintSoloMessage(message)
}
\arguments{
\item{message}{\code{character vector} of length 1 with optional name}
}
\description{
Print text passed in a one-dimensional (named) character vector. Optionally add symbols and specific formatting based on vector element name.
}
