% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{UpdateLocalPackages}
\alias{UpdateLocalPackages}
\title{UpdateLocalPackages}
\usage{
UpdateLocalPackages()
}
\arguments{
\item{Text}{\code{list} List of of named vectors}
}
\description{
Re-Install server and client part of CCPhos during local development
}
\keyword{internal}
