% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CurateData.R
\name{ds.CurateData}
\alias{ds.CurateData}
\title{ds.CurateData}
\usage{
ds.CurateData(
  RawDataSetName = paste0(Module, ".RawDataSet"),
  Module,
  OutputName = paste0(Module, ".CurationOutput"),
  Profile.CurationProcess = "Default",
  Profile.DataRemediation = "Default",
  Profile.Dictionary = "Default",
  Profile.FeatureRequirements = "Default",
  Profile.FeatureTracking = "Default",
  Profile.FuzzyStringMatching = "Default",
  Profile.PrimaryTableCleaning = "Default",
  Profile.RecordSubsumption = "Default",
  Profile.SecondaryTableCleaning = "Default",
  Profile.TableNormalization = "Default",
  Profile.TransformativeExpressions = "Default",
  RunAssignmentChecks = TRUE,
  UnpackCuratedDataSet = TRUE,
  RunSeparately = FALSE,
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{RawDataSetName}{\code{character} - Name of Raw Data Set object (list) on server - Default: '\if{html}{\out{<Module>}}.RawDataSet'}

\item{Module}{\code{string} - Identifying a registered FREDA module (Examples: 'CCP' / 'P21')}

\item{OutputName}{\code{character} - Name of output object to be assigned on server - Default: '\if{html}{\out{<Module>}}.CurationOutput'}

\item{Profile.CurationProcess}{\code{string} - "Default"}

\item{Profile.DataRemediation}{\code{string} - "Default"}

\item{Profile.Dictionary}{\code{string} - "Default"}

\item{Profile.FeatureRequirements}{\code{string} - "Default"}

\item{Profile.FeatureTracking}{\code{string} - "Default"}

\item{Profile.FuzzyStringMatching}{\code{string} - "Default"}

\item{Profile.PrimaryTableCleaning}{\code{string} - "Default"}

\item{Profile.RecordSubsumption}{\code{string} - "Default"}

\item{Profile.SecondaryTableCleaning}{\code{string} - "Default"}

\item{Profile.TableNormalization}{\code{string} - "Default"}

\item{Profile.TransformativeExpressions}{\code{string} - "Default"}

\item{RunAssignmentChecks}{\code{logical} - Indicating whether assignment checks should be performed or omitted for reduced execution time - Default: \code{TRUE}}

\item{UnpackCuratedDataSet}{\code{logical} - Indicating whether the Curated Data Set \code{list} should be unpacked so that tables \code{data.frames} are directly accessible - Default: \code{TRUE}}

\item{RunSeparately}{\code{logical} - Indicating whether \code{CurateDataDS()} should be run separately for each server. This can be done for testing purposes or if timeout issues arise. - Default: \code{FALSE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{flag} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
Messages
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Triggers transformation of Raw Data Set (RDS) into Curated Data Set (CDS) on servers.
}
\details{
Linked to server-side ASSIGN methods \code{CurateDataDS()} and \code{ExtractFromListDS()}
}
\author{
Bastian Reiter
}
