% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.FilterTable.R
\name{ds.FilterTable}
\alias{ds.FilterTable}
\title{ds.FilterTable}
\usage{
ds.FilterTable(
  TableName,
  FilterExpression,
  GroupBy = NULL,
  OutputName,
  DSConnections = NULL
)
}
\arguments{
\item{TableName}{\code{string} - Name of \code{data.frame} on server}

\item{FilterExpression}{\code{string} - \code{dplyr::filter} expression as string}

\item{GroupBy}{\code{string} - Optional \code{dplyr::group_by} expression as string}

\item{OutputName}{\code{string} - Name of resulting \code{data.frame} on server}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Filter a table on server, making use of \code{dplyr} and \code{stringr}.
}
\details{
Linked to server-side \code{ASSIGN} function \code{FilterTableDS()}
}
\author{
Bastian Reiter
}
