% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetCurationReport.R
\name{ds.GetCurationReport}
\alias{ds.GetCurationReport}
\title{ds.GetCurationReport}
\usage{
ds.GetCurationReport(Module = "CCP", DSConnections = NULL)
}
\arguments{
\item{Module}{Optional \code{string} identifying a defined data set and the corresponding meta data (Examples: 'CCP' / 'P21')}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of Curation Reports
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Receive Curation Reports from servers and create cumulated reporting objects.
}
\details{
Linked to server-side \code{AGGREGATE} method \code{GetReportingObjectDS()}
}
\author{
Bastian Reiter
}
