% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetDataSetCheck.R
\name{ds.GetDataSetCheck}
\alias{ds.GetDataSetCheck}
\title{ds.GetDataSetCheck}
\usage{
ds.GetDataSetCheck(
  DataSetName,
  RequiredTableNames = NULL,
  RequiredFeatureNames = NULL,
  EligibleValueSets = NULL,
  Module = "CCP",
  Stage = "Raw",
  DSConnections = NULL
)
}
\arguments{
\item{DataSetName}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet"}

\item{RequiredTableNames}{Optional \code{character vector} - Names of required tables}

\item{RequiredFeatureNames}{Optional \code{list} of \code{character vectors} - Names of required features - Default: \code{names(Table)}}

\item{EligibleValueSets}{Optional \code{list} of \code{character vectors} containing sets of eligible values for corresponding feature.}

\item{Module}{Optional \code{string} identifying a defined data set (Examples: 'CCP' / 'P21')}

\item{Stage}{Optional \code{string} - Indicating transformation stage of addressed data set. This is relevant for which names and values to look up in passed meta data. Options: 'Raw' / 'Curated'}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} containing compiled meta data about data set tables:
\itemize{\item TableStatus
\item TableRowCounts
\item FeatureExistence
\item FeatureTypes
\item NonMissingValueCounts
\item NonMissingValueRates
\item EligibleValueCounts
\item EligibleValueRates}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Checks out a data set (\code{list} of \code{data.frames}) on servers and return a coherent summary across servers. Options:
\enumerate{
\item The data set can be inspected naively without passing any additional information about it
\item Meta data like required table and feature names as well as eligible value sets can be passed explicitly
\item A Module identifier from a list of registered modules can be passed ('CCP' / 'P21' / ...), which leads to meta data being taken from a linked package }
}
\details{
Linked to server-side AGGREGATE method \code{GetDataSetCheckDS()}
}
\author{
Bastian Reiter
}
