% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.PrepareRawData.R
\name{ds.PrepareRawData}
\alias{ds.PrepareRawData}
\title{ds.PrepareRawData}
\usage{
ds.PrepareRawData(
  RawDataSetName,
  Module,
  RDSTableNames,
  FeatureNames.Dictionary = list(),
  FeatureNames.FuzzyStringMatching.Run = FALSE,
  FeatureNames.FuzzyStringMatching.PreferredMethod = "jw",
  FeatureNames.FuzzyStringMatching.Tolerance = 0.2,
  AddIDFeature.Do = FALSE,
  AddIDFeature.IDFeatureName = "ID",
  AddIDFeature.OverwriteExistingIDFeature = FALSE,
  Conversion.IntoCharacter = "None",
  Conversion.DateIntoPOSIXct = NULL,
  CurateFeatureNames = FALSE,
  OutputName = "RDSPreparationOutput",
  RunAssignmentChecks = FALSE,
  PrintMessages = TRUE,
  DSConnections = NULL,
  DS.async = FALSE
)
}
\arguments{
\item{RawDataSetName}{\code{string} - Name of Raw Data Set object (list) on server}

\item{Module}{\code{string} identifying a defined data set and the corresponding meta data needed for feature name harmonization (Examples: 'CCP' / 'P21')}

\item{RDSTableNames}{\code{character vector} - Names of RDS tables}

\item{FeatureNames.Dictionary}{Optional \code{list} containing dictionary data for feature name harmonization (Form: \code{list(Department = c(FAB = "Fachabteilung"))})}

\item{FeatureNames.FuzzyStringMatching.Run}{\code{logical} - Whether to use fuzzy string matching to harmonize raw feature names}

\item{FeatureNames.FuzzyStringMatching.PreferredMethod}{\code{logical} - Selects the method \code{stringdist()} uses (see \code{stringdist} documentation) - Default: 'jw'}

\item{FeatureNames.FuzzyStringMatching.Tolerance}{\code{logical} - Number between 0 and 1 relating to normalized distance between to strings (1 meaning furthest distance, 0 meaning no distance). If a string in 'Vector' is not similar enough to any of the 'EligibleStrings' and its minimal harmonized distance exceeds this number, it is set \code{NA}. - Default: 0.2}

\item{AddIDFeature.Do}{\code{logical} - Whether to add an ID feature (running number) to tables}

\item{AddIDFeature.IDFeatureName}{\code{string} - The name of the new ID feature}

\item{AddIDFeature.OverwriteExistingIDFeature}{\code{logical} - Whether to overwrite an existing feature with the same name}

\item{Conversion.IntoCharacter}{\code{string} - Controls conversion of certain features in data set tables into character type. One of 'None' / 'All' / 'Date'. - Default: 'None'}

\item{Conversion.DateIntoPOSIXct}{\code{list} - Containing character vectors representing date formats used in argument 'tryFormats' in function \code{base::as.POSIXct}. List element should be either '.All' for all date features or specific date feature names. - Default: \code{NULL}}

\item{CurateFeatureNames}{\code{logical} - Indicating whether (after primary harmonization) feature names should be recoded from 'raw' to 'curated' feature names according to Module-specific meta data}

\item{OutputName}{\code{character scalar} - Name of output object to be assigned on server - Default: 'RDSPreparationOutput'}

\item{RunAssignmentChecks}{\code{logical} - Indicating whether assignment checks should be performed or omitted for reduced execution time - Default: \code{FALSE}}

\item{PrintMessages}{\code{logical} - Indicating whether to print messages}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{flag} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{FALSE}}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Trigger basic preparatory transformations on RawDataSet prior to Curation:
\itemize{  \item Feature name harmonization
\item Add ID feature (running number) for tables without primary key feature
\item Type- or format-conversion of features }
}
\details{
Linked to server-side ASSIGN method \code{PrepareRawDataDS()}
}
\author{
Bastian Reiter
}
