% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.PrepareRawData.R
\name{ds.PrepareRawData}
\alias{ds.PrepareRawData}
\title{ds.PrepareRawData}
\usage{
ds.PrepareRawData(
  RawDataSetName,
  Module,
  RDSTableNames,
  FeatureNameDictionary = list(),
  AddIDFeature = list(Do = TRUE, IDFeatureName = "ID", OverwriteExistingIDFeature =
    FALSE),
  CompleteCharacterConversion = FALSE,
  OutputName = "RDSPreparationOutput",
  RunAssignmentChecks = FALSE,
  DSConnections = NULL
)
}
\arguments{
\item{RawDataSetName}{\code{string} - Name of Raw Data Set object (list) on server}

\item{Module}{\code{string} identifying a defined data set and the corresponding meta data needed for feature name harmonization (Examples: 'CCP' / 'P21')}

\item{RDSTableNames}{\code{character vector} - Names of RDS tables}

\item{FeatureNameDictionary}{Optional \code{list} containing dictionary data for feature name harmonization (Form: \code{list(Department = c(FAB = "Fachabteilung"))})}

\item{AddIDFeature}{\code{list} containing parameters about adding an ID feature to tables:
\itemize{ \item Do (\code{logical}) - Whether to add an ID feature (running number)
\item IDFeatureName (\code{string})
\item OverwriteExistingIDFeature (\code{logical}) - Whether to overwrite an existing feature with the same name }}

\item{CompleteCharacterConversion}{\code{logical} - Indicating whether to convert all features in data set tables to character type}

\item{OutputName}{\code{character scalar} - Name of output object to be assigned on server - Default: 'RDSPreparationOutput'}

\item{RunAssignmentChecks}{\code{logical} Indicating whether assignment checks should be performed or omitted for reduced execution time - Default: \code{FALSE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Trigger basic preparatory transformations on RawDataSet prior to Curation:
\itemize{  \item Optionally convert all columns to character type
\item Add ID feature (running number) for tables without primary key feature
\item Try to harmonize feature names employing Fuzzy String Matching and Dictionary look-up }
}
\details{
Linked to server-side ASSIGN method \code{PrepareRawDataDS()}
}
\author{
Bastian Reiter
}
