% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreFeature.R
\name{ExploreFeature}
\alias{ExploreFeature}
\title{ExploreFeature}
\usage{
ExploreFeature(TableName, FeatureName, DSConnections = NULL, ...)
}
\arguments{
\item{TableName}{\code{string} - Name of the table containing the feature of concern}

\item{FeatureName}{\code{string} - Name of feature}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
\code{list}
\itemize{ \item FeatureInfo
\item Statistics }
}
\description{
Get characterizing statistics about a feature of arbitrary data type.
}
\author{
Bastian Reiter
}
