% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetServerWorkspaceInfo.R
\name{GetServerWorkspaceInfo}
\alias{GetServerWorkspaceInfo}
\title{GetServerWorkspaceInfo}
\usage{
GetServerWorkspaceInfo(DSConnections = NULL)
}
\arguments{
\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} containing overview and details of server-side workspace objects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Check which objects live in server-side R sessions and collect meta data about them.
}
\author{
Bastian Reiter
}
