% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBoxPlot.R
\name{MakeBoxPlot}
\alias{MakeBoxPlot}
\title{MakeBoxPlot}
\usage{
MakeBoxPlot(
  SampleStatistics,
  AxisLimits_y = c(NA_integer_, NA_integer_),
  AxisTitle_x = "",
  AxisTitle_y = "",
  TickLabelWidth_x = 10,
  Decimals = 0,
  ggTheme = function(...) dsFredaClient::ggTheme(...),
  ThemeArguments = list(),
  FillPalette = NULL,
  ...
)
}
\arguments{
\item{SampleStatistics}{\code{data.frame} - Sample statistics as generated by \code{\link{ds.GetSampleStatistics}}}

\item{AxisLimits_y}{\code{integer} containing two integer values - Default: Auto y axis limits}

\item{AxisTitle_x}{\code{string}}

\item{AxisTitle_y}{\code{string}}

\item{TickLabelWidth_x}{\code{integer}}

\item{Decimals}{Number of decimals of y axis number format}

\item{ggTheme}{Pass custom theme}

\item{ThemeArguments}{Pass custom theme arguments}

\item{FillPalette}{\code{character vector}}

\item{...}{}
}
\value{
a \code{ggplot} object
}
\description{
Make one or more box plots on the basis of sample statistics as generated by \code{\link{ds.GetSampleStatistics}}. Makes use of ggplot2.
}
\author{
Bastian Reiter
}
