% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPerformanceTest.R
\name{RunPerformanceTest}
\alias{RunPerformanceTest}
\title{RunPerformanceTest}
\usage{
RunPerformanceTest(
  ServerSpecifications,
  Module = "CCP",
  ScenarioA.Runs = 1,
  ScenarioA.ServerCount = nrow(ServerSpecifications),
  ScenarioA.SampleSizes = c(125, 250, 500, 1000, 2000),
  ScenarioA.SampleSizes.Shuffle = TRUE,
  ScenarioB.Runs = 1,
  ScenarioB.SequenceServerCount = 1:nrow(ServerSpecifications),
  ScenarioB.SampleSize = 500,
  CreateLogFiles = TRUE,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} - Same \code{data.frame} used for login. Used here only for acquisition of server-specific project names (in case they are differing) - Default: \code{NULL} for virtual project}

\item{Module}{\code{string} - Identifying a registered FREDA module (Examples: 'CCP' / 'P21')}

\item{ScenarioA.Runs}{\code{integer} - Number of runs through performance test scenario A}

\item{ScenarioA.SampleSizes}{\code{integer}}

\item{ScenarioA.SampleSizes.Shuffle}{\code{logical} - Whether to shuffle sample size numbers between runs in order to mitigate possible sequence position effects - Default: \code{TRUE}}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
A \code{list} of report objects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Perform multiple data preprocessing runs of FREDA functionality while tracking performance measures.
}
\author{
Bastian Reiter
}
